rainbow seeds
i’m tired now, but tomorrow i’ll add the tree, then the bifrost-colored
stuff will be ready to push

the wood will NOT be shimmerwood. it will be iridescent, like
everything else, as shimmerwood is made of magic and wood, while
Iridescent Wood is made of magic absorbed through natural processes in
a magically devoid sapling.